[chore #161341913] Create index.js file to consolidate model imports
